IMTA-13449: Reverting type of number of animals field back to integer

See merge request imports/imports-notification-schema!300

+++ notification-schema-java/src/main/java/uk/gov/defra/tracesx/notificationschema/representation/Commodities.java
@@ -84,7 +84,7 @@ public class Commodities {
       message =
           "{uk.gov.defra.tracesx.notificationschema.representation.partone.commodities"
               + ".numberofanimals.min.message}")
-  private Long numberOfAnimals = null;
+  private Integer numberOfAnimals = null;
 
   @NotNull(
       groups = NotificationLowRiskFieldValidation.class,
